Reorganize barostat tests
* move molecular ideal gas test into slow_tests
* rename remaining tests
